+++ app/src/test/java/com/rikvanvelzen/codingtest/ui/utils/StringFormatUtilTest.kt
@@ -6,46 +6,31 @@
 
 package com.rikvanvelzen.codingtest.ui.utils
 
+import org.hamcrest.CoreMatchers.`is`
+import org.junit.Assert.assertThat
 import org.junit.Before
 import org.junit.Test
-import org.mockito.ArgumentCaptor
+import org.junit.runner.RunWith
 import org.mockito.junit.MockitoJUnitRunner
 
-import org.junit.Assert
-import org.junit.runner.RunWith
-import org.mockito.Mock
-
-import org.hamcrest.CoreMatchers.`is`
-import org.hamcrest.CoreMatchers.nullValue
-import org.junit.Assert.assertThat
-import org.mockito.Mockito.*
-
 /*****************************************************
-* <unitOfWork>_<stateUnderTest>_<expectedBehavior>
-*
-* - camelCased
-* - single underscore separated parts
-*
-* - unitOfWork         =>  most cases a method name. Can also be general operation or flow
-* - expectedBehavior   =>  result/behavior when the test completes. Can be retun value,
-*                          thrown exception, external interaction, etc.
-*
-* Nullability: - Assumptions in all test => method arguments and return values are non-null (?!) by default
-*              - Nullable values will explicitely be annotated with @Null
-*              - No need to test nulls in most(!) cases
-* Boundary Conditions: make sure all boundary conditions are tested
-*
-******************************************************/
+ * <unitOfWork>_<stateUnderTest>_<expectedBehavior>
+ *
+ * - camelCased
+ * - single underscore separated parts
+ *
+ * - unitOfWork         =>  most cases a method name. Can also be general operation or flow
+ * - expectedBehavior   =>  result/behavior when the test completes. Can be retun value,
+ *                          thrown exception, external interaction, etc.
+ *
+ * Nullability: - Assumptions in all test => method arguments and return values are non-null (?!) by default
+ *              - Nullable values will explicitly be annotated with @Null
+ *              - No need to test nulls in most(!) cases
+ * Boundary Conditions: make sure all boundary conditions are tested
+ *
+ ******************************************************/
 @RunWith(MockitoJUnitRunner::class)
 class StringFormatUtilTest {
-
-    // region constants
-
-    // end region constants
-
-    // region helper fields
-
-    // end region helper fields
 
     private lateinit var SUT: StringFormatUtil
 
@@ -57,7 +42,7 @@ class StringFormatUtilTest {
     @Test
     fun getFormattedExchangeRate_nullValuePassed_emptyStringReturned() {
         // arrange
-         val rate = null
+        val rate = null
         // act
         val result = SUT.getFormattedExchangeRate(rate)
         // assert
@@ -67,7 +52,7 @@ class StringFormatUtilTest {
     @Test
     fun getFormattedExchangeRate_noDecimalPartOnFloatValue_intValueAsStringReturned() {
         // arrange
-        val rate = 9.toDouble()
+        val rate = 9.toBigDecimal()
         // act
         val result = SUT.getFormattedExchangeRate(rate)
         // assert
@@ -77,7 +62,7 @@ class StringFormatUtilTest {
     @Test
     fun getFormattedExchangeRate_decimalPartsValueIsZero_intValueAsStringReturned() {
         // arrange
-        val rate = 88.0000
+        val rate = 88.0000.toBigDecimal()
         // act
         val result = SUT.getFormattedExchangeRate(rate)
         // assert
@@ -87,7 +72,7 @@ class StringFormatUtilTest {
     @Test
     fun getFormattedExchangeRate_singleDecimalValue_valueWithZeroAsSecondDecimalReturned() {
         // arrange
-        val rate = 88.8
+        val rate = 88.8.toBigDecimal()
         // act
         val result = SUT.getFormattedExchangeRate(rate)
         // assert
@@ -97,7 +82,7 @@ class StringFormatUtilTest {
     @Test
     fun getFormattedExchangeRate_twoDecimalsWithNotZeroValue_valueAsStringReturned() {
         // arrange
-        val rate = 88.80
+        val rate = 88.80.toBigDecimal()
         // act
         val result = SUT.getFormattedExchangeRate(rate)
         // assert
@@ -107,7 +92,7 @@ class StringFormatUtilTest {
     @Test
     fun getFormattedExchangeRate_thirdDecimalFiveOrHigher_valueWithTwoDecimalsRoundedUpAsStringReturned() {
         // arrange
-        val rate = 88.805000
+        val rate = 88.805000.toBigDecimal()
         // act
         val result = SUT.getFormattedExchangeRate(rate)
         // assert
@@ -117,20 +102,10 @@ class StringFormatUtilTest {
     @Test
     fun getFormattedExchangeRate_thirdDecimalBelowFive_valueWithTwoDecimalsNotRoundedUpAsStringReturned() {
         // arrange
-        val rate = 88.804999999
+        val rate = 88.804999999.toBigDecimal()
         // act
         val result = SUT.getFormattedExchangeRate(rate)
         // assert
         assertThat(result, `is`("88.80"))
     }
-
-
-
-    // region helper methods
-
-    // end region helper methods
-
-    // region helper classes
-
-    // end region helper classes
 }
